+++ core/server-ng/src/bootstrap.rs
@@ -18,65 +18,24 @@
  */
 
 use crate::config_writer::write_current_config;
-use crate::login_register::LoginRegisterError;
+use crate::dispatch::{
+    make_client_request_handler, make_deferred_client_request_handler,
+    make_deferred_replica_message_handler, make_list_clients_handler, make_metadata_submit_handler,
+};
 use crate::server_error::{ServerNgError, ShardJoinFailure, ShardJoinFailureKind};
-use crate::session_manager::{ClientRecord, SessionManager};
-use bytes::Bytes;
+use crate::session_manager::SessionManager;
 use configs::server_ng::ServerNgConfig;
 use configs::sharding::{
     INBOX_CAPACITY_MAX, SHUTDOWN_DRAIN_TIMEOUT_MAX, SHUTDOWN_POLL_INTERVAL_MAX,
 };
-use consensus::{LocalPipeline, MetadataHandle, PartitionsHandle, Sequencer, VsrConsensus};
+use consensus::{LocalPipeline, PartitionsHandle, Sequencer, VsrConsensus};
 // `try_send` / `try_recv` resolve through these traits on `MAsyncTx` /
 // `MAsyncRx`; the metadata-handoff loops below depend on the
 // non-blocking variants for cancel-safe shutdown polling.
 use crossfire::{AsyncRxTrait, AsyncTxTrait};
-use iggy_binary_protocol::codes::{
-    GET_CLIENT_CODE, GET_CLIENTS_CODE, GET_CLUSTER_METADATA_CODE, GET_ME_CODE, GET_STATS_CODE,
-    GET_STREAM_CODE, GET_STREAMS_CODE, GET_TOPIC_CODE, GET_TOPICS_CODE, GET_USER_CODE,
-    GET_USERS_CODE, PING_CODE, POLL_MESSAGES_CODE,
-};
-use iggy_binary_protocol::requests::consumer_offsets::{
-    DeleteConsumerOffset2Request, DeleteConsumerOffsetRequest, StoreConsumerOffset2Request,
-    StoreConsumerOffsetRequest,
-};
-use iggy_binary_protocol::requests::messages::{PollMessagesRequest, SendMessagesHeader};
-use iggy_binary_protocol::requests::personal_access_tokens::{
-    CreatePersonalAccessTokenRequest as WireCreatePersonalAccessTokenRequest,
-    DeletePersonalAccessTokenRequest as WireDeletePersonalAccessTokenRequest,
-};
-use iggy_binary_protocol::requests::segments::DeleteSegmentsRequest;
-use iggy_binary_protocol::requests::streams::{GetStreamRequest, GetStreamsRequest};
-use iggy_binary_protocol::requests::system::get_client::GetClientRequest;
-use iggy_binary_protocol::requests::topics::{GetTopicRequest, GetTopicsRequest};
-use iggy_binary_protocol::requests::users::{
-    GetUserRequest, LoginRegisterRequest, LoginRegisterWithPatRequest,
-};
-use iggy_binary_protocol::responses::clients::client_response::ClientResponse;
-use iggy_binary_protocol::responses::clients::get_client::ClientDetailsResponse;
-use iggy_binary_protocol::responses::clients::get_clients::GetClientsResponse;
-use iggy_binary_protocol::responses::personal_access_tokens::RawPersonalAccessTokenResponse;
-use iggy_binary_protocol::responses::streams::StreamResponse;
-use iggy_binary_protocol::responses::streams::get_stream::{
-    GetStreamResponse, TopicHeader as StreamTopicHeader,
-};
-use iggy_binary_protocol::responses::streams::get_streams::GetStreamsResponse;
-use iggy_binary_protocol::responses::system::get_cluster_metadata::{
-    ClusterMetadataResponse, ClusterNodeResponse,
-};
-use iggy_binary_protocol::responses::system::get_stats::StatsResponse;
-use iggy_binary_protocol::responses::topics::get_topic::{GetTopicResponse, PartitionResponse};
-use iggy_binary_protocol::responses::topics::get_topics::GetTopicsResponse;
-use iggy_binary_protocol::responses::users::get_user::UserDetailsResponse;
-use iggy_binary_protocol::responses::users::get_users::GetUsersResponse;
-use iggy_binary_protocol::responses::users::user_response::UserResponse;
-use iggy_binary_protocol::{
-    Command2, EvictionReason, GenericHeader, Operation, ReplyHeader, RequestHeader, WireDecode,
-    WireEncode, WireIdentifier, WireName, WirePartitioning,
-};
 use iggy_common::{
-    ConsumerGroupOffsets, ConsumerOffsets, IggyByteSize, IggyError, IggyTimestamp, PartitionStats,
-    PersonalAccessToken, TopicStats, UserStatus, variadic,
+    ConsumerGroupOffsets, ConsumerOffsets, IggyByteSize, IggyError, PartitionStats, TopicStats,
+    variadic,
 };
 use journal::Journal;
 use journal::prepare_journal::PrepareJournal;
@@ -84,54 +43,45 @@ use message_bus::client_listener::{self, RequestHandler};
 use message_bus::installer;
 use message_bus::installer::conn_info::{ClientConnMeta, ClientTransportKind};
 use message_bus::replica::io as replica_io;
-use message_bus::replica::listener::{self as replica_listener, MessageHandler};
+use message_bus::replica::listener::{self as replica_listener};
 use message_bus::transports::quic::server_config_with_cert;
 use message_bus::transports::tls::{
     TlsServerCredentials, install_default_crypto_provider, load_pem, self_signed_for_loopback,
 };
 use message_bus::{
     AcceptedClientFn, AcceptedQuicClientFn, AcceptedReplicaFn, AcceptedTlsClientFn,
-    AcceptedWsClientFn, IggyMessageBus, MessageBus, ReplicaOwnerTable, connector,
+    AcceptedWsClientFn, IggyMessageBus, ReplicaOwnerTable, connector,
 };
 use metadata::IggyMetadata;
 use metadata::MuxStateMachine;
-use metadata::impls::metadata::{IggySnapshot, RegisterSubmitError, StreamsFrontend};
+use metadata::impls::metadata::{IggySnapshot, StreamsFrontend};
 use metadata::impls::recovery::recover;
 use metadata::stm::consumer_group::ConsumerGroups;
 use metadata::stm::mux::WithFactory;
 use metadata::stm::snapshot::Snapshot;
 use metadata::stm::stream::{Partition, Streams};
 use metadata::stm::user::Users;
-use metadata::stm::user::{
-    CreatePersonalAccessTokenRequest as ReplicatedCreatePersonalAccessTokenRequest,
-    DeletePersonalAccessTokenRequest as ReplicatedDeletePersonalAccessTokenRequest,
-};
 use partitions::{
     IggyIndexWriter, IggyPartition, IggyPartitions, MessagesWriter, PartitionsConfig, Segment,
 };
-use server_common::Message;
 use server_common::bootstrap::create_directories;
 use server_common::executor::create_shard_executor;
 use server_common::sharding::{IggyNamespace, LocalIdx, PartitionLocation, ShardId};
 // TODO: decouple bootstrap/storage helpers and logging from the `server` crate.
-use secrecy::ExposeSecret;
 use server::log::logger::Logging;
 use server::shard_allocator::{ShardAllocator, ShardInfo};
 use server::streaming::partitions::storage::{load_consumer_group_offsets, load_consumer_offsets};
 use server::streaming::segments::storage::create_segment_storage;
 use server::streaming::users::user::User as LegacyUser;
-use server::streaming::utils::crypto;
 use server::{IGGY_ROOT_PASSWORD_ENV, IGGY_ROOT_USERNAME_ENV};
 use shard::builder::IggyShardBuilder;
 use shard::metrics::ShardMetrics;
 use shard::shards_table::{PapayaShardsTable, calculate_shard_assignment};
 use shard::{
-    ConnectedClientInfo, CoordinatorConfig, IggyShard, ListClientsHandler,
-    PartitionConsensusConfig, Receiver as ShardReceiver, ShardFrame, ShardIdentity, TaggedSender,
-    channel, shard_mesh_channels,
+    CoordinatorConfig, IggyShard, PartitionConsensusConfig, Receiver as ShardReceiver, ShardFrame,
+    ShardIdentity, TaggedSender, channel, shard_mesh_channels,
 };
 use std::cell::RefCell;
-use std::collections::{HashMap, HashSet, VecDeque};
 use std::env;
 use std::net::{IpAddr, SocketAddr};
 use std::path::{Path, PathBuf};
@@ -153,7 +103,7 @@ type ServerNgMuxStateMachine = MuxStateMachine<variadic!(Users, Streams, Consume
 /// [`ServerNgMuxStateMachine`] on its own runtime, skipping the WAL.
 type ServerNgMetadataBundle = <variadic!(Users, Streams, ConsumerGroups) as WithFactory>::Bundle;
 
-type ServerNgMetadata = IggyMetadata<
+pub(crate) type ServerNgMetadata = IggyMetadata<
     VsrConsensus<Rc<IggyMessageBus>>,
     PrepareJournal,
     IggySnapshot,
@@ -167,7 +117,7 @@ pub type ServerNgShard = IggyShard<
     PapayaShardsTable,
 >;
 
-type ServerNgShardHandle = Rc<RefCell<Option<Weak<ServerNgShard>>>>;
+pub(crate) type ServerNgShardHandle = Rc<RefCell<Option<Weak<ServerNgShard>>>>;
 
 /// Result of a multi-shard bootstrap.
 ///
@@ -2086,1653 +2036,6 @@ async fn start_manual_runtime(
     Ok(())
 }
 
-type ClientRequestQueues = Rc<RefCell<HashMap<u128, VecDeque<Message<GenericHeader>>>>>;
-type ActiveClientRequests = Rc<RefCell<HashSet<u128>>>;
-
-fn make_client_request_handler(
-    shard: &Rc<ServerNgShard>,
-    sessions: &Rc<RefCell<SessionManager>>,
-) -> RequestHandler {
-    let shard = Rc::clone(shard);
-    let sessions = Rc::clone(sessions);
-    let queues: ClientRequestQueues = Rc::new(RefCell::new(HashMap::new()));
-    let active: ActiveClientRequests = Rc::new(RefCell::new(HashSet::new()));
-    let sessions_for_disconnect = Rc::clone(&sessions);
-    let shard_for_disconnect = Rc::clone(&shard);
-    shard
-        .bus
-        .set_client_connection_lost_fn(Rc::new(move |client_id| {
-            if let Some(vsr_client_id) = sessions_for_disconnect
-                .borrow_mut()
-                .remove_connection(client_id)
-            {
-                // Local-table reclaim happens here; then spawn a detached
-                // task to submit a cluster-replicated `Logout` so peer
-                // replicas drop the slot too. Without this, an
-                // already-committed `Register` lingers on every backup
-                // until they evict the client themselves -- a
-                // local-vs-cluster divergence observable across view
-                // changes. `submit_logout_in_process` requires shard 0
-                // consensus, which this handler always runs on.
-                shard_for_disconnect
-                    .plane
-                    .metadata()
-                    .remove_client_session(vsr_client_id);
-                let shard = Rc::clone(&shard_for_disconnect);
-                compio::runtime::spawn(async move {
-                    // request_id 0 + session 0: synthetic disconnect
-                    // logout; `submit_logout_in_process` short-circuits to
-                    // current commit when no session entry exists, so a
-                    // race with the explicit logout path is a no-op.
-                    if let Err(error) = shard
-                        .plane
-                        .metadata()
-                        .submit_logout_in_process(vsr_client_id, 0, 0)
-                        .await
-                    {
-                        warn!(
-                            vsr_client_id,
-                            ?error,
-                            "disconnect logout submit failed; peer slots may linger until eviction"
-                        );
-                    }
-                })
-                .detach();
-            }
-        }));
-    Rc::new(move |client_id, message| {
-        enqueue_client_request(
-            Rc::clone(&shard),
-            Rc::clone(&sessions),
-            Rc::clone(&queues),
-            Rc::clone(&active),
-            client_id,
-            message,
-        );
-    })
-}
-
-/// Build the per-shard [`ListClientsHandler`]: on a `ListClients`
-/// broadcast, serialize this shard's locally-homed connected clients from
-/// its `SessionManager` and push them back over the reply sender. The
-/// aggregation across all shards happens in
-/// [`shard::IggyShard::list_all_clients`].
-fn make_list_clients_handler(sessions: &Rc<RefCell<SessionManager>>) -> ListClientsHandler {
-    let sessions = Rc::clone(sessions);
-    Rc::new(move |reply| {
-        let clients: Vec<ConnectedClientInfo> = sessions
-            .borrow()
-            .iter_clients()
-            .map(|record| ConnectedClientInfo {
-                client_id: record.client_id,
-                user_id: record.user_id,
-                transport: record.transport,
-                address: record.address,
-            })
-            .collect();
-        // Best-effort: the gather side bounds itself by count + timeout, so
-        // a dropped reply (receiver gone) just means this shard is omitted.
-        let _ = reply.try_send(clients);
-    })
-}
-
-fn make_deferred_replica_message_handler(shard_handle: &ServerNgShardHandle) -> MessageHandler {
-    let shard_handle = Rc::clone(shard_handle);
-    Rc::new(move |_replica_id, message| {
-        if let Some(shard) = upgrade_shard_handle(&shard_handle) {
-            shard.dispatch(message);
-        }
-    })
-}
-
-fn make_deferred_client_request_handler(
-    bus: &Rc<IggyMessageBus>,
-    shard_handle: &ServerNgShardHandle,
-    sessions: &Rc<RefCell<SessionManager>>,
-) -> RequestHandler {
-    let shard_handle = Rc::clone(shard_handle);
-    let sessions = Rc::clone(sessions);
-    let queues: ClientRequestQueues = Rc::new(RefCell::new(HashMap::new()));
-    let active: ActiveClientRequests = Rc::new(RefCell::new(HashSet::new()));
-    let sessions_for_disconnect = Rc::clone(&sessions);
-    let shard_handle_for_disconnect = Rc::clone(&shard_handle);
-    bus.set_client_connection_lost_fn(Rc::new(move |client_id| {
-        if let Some(vsr_client_id) = sessions_for_disconnect
-            .borrow_mut()
-            .remove_connection(client_id)
-            && let Some(shard) = upgrade_shard_handle(&shard_handle_for_disconnect)
-        {
-            // TODO(vsr): this handler runs on shards > 0 which have no
-            // consensus; `submit_logout_in_process` would panic. Local
-            // table reclaim only -- bounded local-vs-cluster divergence
-            // window until `evict_oldest` reclaims the peer slot. Fix by
-            // routing a synthetic Logout request through the inter-shard
-            // router to shard 0.
-            shard.plane.metadata().remove_client_session(vsr_client_id);
-        }
-    }));
-    Rc::new(move |client_id, message| {
-        let shard_handle = Rc::clone(&shard_handle);
-        let sessions = Rc::clone(&sessions);
-        let queues = Rc::clone(&queues);
-        let active = Rc::clone(&active);
-        queues
-            .borrow_mut()
-            .entry(client_id)
-            .or_default()
-            .push_back(message);
-        if !active.borrow_mut().insert(client_id) {
-            return;
-        }
-        compio::runtime::spawn(async move {
-            let Some(shard) = upgrade_shard_handle(&shard_handle) else {
-                active.borrow_mut().remove(&client_id);
-                return;
-            };
-            drain_client_requests(shard, sessions, queues, active, client_id).await;
-        })
-        .detach();
-    })
-}
-
-/// Handler shard 0 runs for an inbound [`shard::MetadataSubmit`]: a peer
-/// shard has verified credentials and owns the session locally, and asks
-/// shard 0 (the metadata consensus owner) to run only the consensus
-/// proposal. Spawns a task so the awaiting peer is woken once the op
-/// commits; replies `None` on transient submit failure so the peer never
-/// blocks forever.
-fn make_metadata_submit_handler(
-    shard_handle: &ServerNgShardHandle,
-) -> shard::MetadataSubmitHandler {
-    let shard_handle = Rc::clone(shard_handle);
-    Rc::new(move |submit| {
-        let shard_handle = Rc::clone(&shard_handle);
-        compio::runtime::spawn(async move {
-            let Some(shard) = upgrade_shard_handle(&shard_handle) else {
-                return;
-            };
-            match submit {
-                shard::MetadataSubmit::Register {
-                    vsr_client_id,
-                    reply,
-                } => {
-                    let session = shard
-                        .plane
-                        .metadata()
-                        .submit_register_in_process(vsr_client_id)
-                        .await
-                        .ok();
-                    let _ = reply.try_send(session);
-                }
-                shard::MetadataSubmit::Logout {
-                    vsr_client_id,
-                    session,
-                    request,
-                    reply,
-                } => {
-                    let commit = shard
-                        .plane
-                        .metadata()
-                        .submit_logout_in_process(vsr_client_id, session, request)
-                        .await
-                        .ok();
-                    let _ = reply.try_send(commit);
-                }
-                shard::MetadataSubmit::ClientRequest { request, reply } => {
-                    let committed = match request.try_into_typed::<RequestHeader>() {
-                        Ok(typed) => shard
-                            .plane
-                            .metadata()
-                            .submit_request_in_process(typed)
-                            .await
-                            .ok()
-                            .map(server_common::Message::into_generic),
-                        Err(error) => {
-                            warn!(?error, "ClientRequest submit: undecodable request header");
-                            None
-                        }
-                    };
-                    let _ = reply.try_send(committed);
-                }
-            }
-        })
-        .detach();
-    })
-}
-
-fn enqueue_client_request(
-    shard: Rc<ServerNgShard>,
-    sessions: Rc<RefCell<SessionManager>>,
-    queues: ClientRequestQueues,
-    active: ActiveClientRequests,
-    client_id: u128,
-    message: Message<GenericHeader>,
-) {
-    queues
-        .borrow_mut()
-        .entry(client_id)
-        .or_default()
-        .push_back(message);
-    if !active.borrow_mut().insert(client_id) {
-        return;
-    }
-
-    compio::runtime::spawn(async move {
-        drain_client_requests(shard, sessions, queues, active, client_id).await;
-    })
-    .detach();
-}
-
-#[allow(clippy::future_not_send)]
-async fn drain_client_requests(
-    shard: Rc<ServerNgShard>,
-    sessions: Rc<RefCell<SessionManager>>,
-    queues: ClientRequestQueues,
-    active: ActiveClientRequests,
-    client_id: u128,
-) {
-    loop {
-        let Some(message) = pop_next_client_request(&queues, &active, client_id) else {
-            return;
-        };
-        handle_client_request(&shard, &sessions, client_id, message).await;
-    }
-}
-
-fn pop_next_client_request(
-    queues: &ClientRequestQueues,
-    active: &ActiveClientRequests,
-    client_id: u128,
-) -> Option<Message<GenericHeader>> {
-    let mut queues = queues.borrow_mut();
-    let Some(queue) = queues.get_mut(&client_id) else {
-        active.borrow_mut().remove(&client_id);
-        return None;
-    };
-    let message = queue.pop_front();
-    if queue.is_empty() {
-        queues.remove(&client_id);
-    }
-    if message.is_none() {
-        active.borrow_mut().remove(&client_id);
-    }
-    message
-}
-
-#[allow(clippy::future_not_send, clippy::too_many_lines)]
-async fn handle_client_request(
-    shard: &Rc<ServerNgShard>,
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-    message: Message<iggy_binary_protocol::GenericHeader>,
-) {
-    let request = match message.try_into_typed::<RequestHeader>() {
-        Ok(request) => request,
-        Err(error) => {
-            warn!(
-                transport_client_id,
-                error = %error,
-                "dropping client request with invalid header"
-            );
-            return;
-        }
-    };
-
-    ensure_transport_connection(shard, sessions, transport_client_id);
-
-    let header = *request.header();
-    if header.operation == Operation::NonReplicated {
-        // Auth bypass guard: only `PING` and `GET_CLUSTER_METADATA` are
-        // legitimately pre-auth (liveness probe + connection bootstrap
-        // metadata). Every other non-replicated code (`GET_STREAM*`,
-        // `GET_TOPIC*`, `GET_STATS`, `POLL_MESSAGES`) reads live state and
-        // MUST go through Register first, since server-ng has no
-        // per-resource authz layer.
-        let nr_code = u32::from_le_bytes(
-            request.header().reserved[..4]
-                .try_into()
-                .unwrap_or_default(),
-        );
-        let allowed_pre_auth = matches!(nr_code, PING_CODE | GET_CLUSTER_METADATA_CODE);
-        if !allowed_pre_auth && sessions.borrow().get_session(transport_client_id).is_none() {
-            warn!(
-                transport_client_id,
-                code = nr_code,
-                "dropping pre-auth non-replicated read"
-            );
-            return;
-        }
-        handle_non_replicated_request(shard, sessions, transport_client_id, request).await;
-        return;
-    }
-
-    if header.operation == Operation::Register && header.session == 0 && header.request == 0 {
-        handle_login_register_request(shard, sessions, transport_client_id, request).await;
-        return;
-    }
-
-    if header.operation == Operation::Logout {
-        handle_logout_request(shard, sessions, transport_client_id, request).await;
-        return;
-    }
-
-    let bound = sessions.borrow().get_session(transport_client_id);
-    if bound.is_none() {
-        // Replicated request on an unbound transport. Without this short-
-        // circuit, the rewrite below overwrites `header.client` with
-        // `transport_client_id` and dispatches; the request_preflight then
-        // rejects with `NoSession`/`SessionMismatch` and the failure either
-        // disappears silently or emits an Eviction the SDK previously
-        // could not decode. Either way the SDK blocked until socket
-        // timeout. Emit an empty Reply so the SDK fails fast: the typed
-        // decoder downstream rejects the empty body with `InvalidCommand`
-        // instead of hanging.
-        let commit = current_metadata_commit(shard);
-        let reply = build_empty_reply(&header, transport_client_id, 0, commit);
-        if let Err(error) = shard
-            .bus
-            .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-            .await
-        {
-            warn!(
-                transport_client_id,
-                error = %error,
-                operation = ?header.operation,
-                "failed to surface unbound-session reply"
-            );
-        } else {
-            warn!(
-                transport_client_id,
-                operation = ?header.operation,
-                "dropping replicated request from unbound transport; replied empty"
-            );
-        }
-        return;
-    }
-
-    let resolved_namespace = if header.operation.is_partition() {
-        match resolve_partition_request_namespace(shard, header.operation, request_body(&request)) {
-            Ok(namespace) => Some(namespace),
-            Err(error) => {
-                warn!(
-                    transport_client_id,
-                    error = %error,
-                    operation = ?header.operation,
-                    "dropping partition request with unresolved namespace"
-                );
-                return;
-            }
-        }
-    } else {
-        None
-    };
-    let request = request.transmute_header(|header, new_header: &mut RequestHeader| {
-        *new_header = header;
-        new_header.client = transport_client_id;
-        if let Some(namespace) = resolved_namespace {
-            new_header.namespace = namespace;
-        }
-        if let Some((bound_client_id, bound_session)) = bound {
-            new_header.client = bound_client_id;
-            new_header.session = bound_session;
-        }
-    });
-    let (request, raw_pat_token) =
-        match maybe_rewrite_pat_request(sessions, transport_client_id, request) {
-            Ok(rewritten) => rewritten,
-            Err(error) => {
-                warn!(
-                    transport_client_id,
-                    error = %error,
-                    operation = ?header.operation,
-                    "dropping request with invalid PAT replication context"
-                );
-                return;
-            }
-        };
-    let request_header = *request.header();
-    // Replicated request: run consensus on the metadata owner (shard 0) and
-    // bring the committed reply back here. This shard owns the connection,
-    // so it writes the reply to the socket via the transport client id --
-    // shard 0 can't route by the consensus client id (no home-shard bits).
-    match submit_client_request_on_owner(shard, request).await {
-        Some(reply) => {
-            // The raw PAT token never enters consensus (it is non-deterministic
-            // and secret), so the committed reply body is empty. Substitute the
-            // raw-token response here, on the minting client's home shard, using
-            // the confirmed commit position from the committed reply.
-            let reply = match build_raw_pat_reply(&request_header, reply, raw_pat_token) {
-                Ok(reply) => reply,
-                Err(error) => {
-                    warn!(
-                        transport_client_id,
-                        error = %error,
-                        "failed to build raw PAT reply"
-                    );
-                    return;
-                }
-            };
-            if let Err(error) = shard
-                .bus
-                .send_to_client(transport_client_id, reply.into_frozen())
-                .await
-            {
-                warn!(
-                    transport_client_id,
-                    error = %error,
-                    operation = ?header.operation,
-                    "failed to deliver committed reply to client"
-                );
-            }
-        }
-        None => {
-            // Transient submit failure (not primary / not caught up / dedup
-            // absorbed). Stay silent; the SDK read-timeout replays.
-            warn!(
-                transport_client_id,
-                operation = ?header.operation,
-                "replicated request not committed (transient); client will replay"
-            );
-        }
-    }
-}
-
-#[allow(clippy::future_not_send)]
-async fn handle_non_replicated_request(
-    shard: &Rc<ServerNgShard>,
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-    request: Message<RequestHeader>,
-) {
-    const CODE_RANGE: std::ops::Range<usize> = 0..4;
-    let code = u32::from_le_bytes(request.header().reserved[CODE_RANGE].try_into().unwrap());
-    match code {
-        PING_CODE => {
-            let commit = current_metadata_commit(shard);
-            let reply = build_empty_reply(
-                request.header(),
-                request.header().client,
-                request.header().session,
-                commit,
-            );
-            if let Err(error) = shard
-                .bus
-                .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-                .await
-            {
-                warn!(
-                    transport_client_id,
-                    error = %error,
-                    "failed to send non-replicated ping reply"
-                );
-            }
-        }
-        GET_ME_CODE => {
-            // `get_me` reports the requesting connection's own identity,
-            // sourced from this shard's `SessionManager` (not `IggyMetadata`),
-            // so it is built here rather than in
-            // `build_non_replicated_response`.
-            let response = build_get_me_response(sessions, transport_client_id);
-            send_non_replicated_bytes(
-                shard,
-                &request,
-                transport_client_id,
-                response.to_bytes(),
-                "get_me",
-            )
-            .await;
-        }
-        GET_CLIENTS_CODE => {
-            // Shared-nothing: each shard knows only its own connections, so
-            // gather across all shards (scatter-gather over the mesh).
-            let infos = shard.list_all_clients().await;
-            let response = GetClientsResponse {
-                clients: infos.iter().map(connected_client_to_response).collect(),
-            };
-            send_non_replicated_bytes(
-                shard,
-                &request,
-                transport_client_id,
-                response.to_bytes(),
-                "get_clients",
-            )
-            .await;
-        }
-        GET_CLIENT_CODE => {
-            // No reverse map from the wire u32 id to a u128 transport id /
-            // home shard (the u32 is just the seq tail), so gather all and
-            // filter -- same fan-out as `get_clients`.
-            let target = GetClientRequest::decode_from(request_body(&request))
-                .ok()
-                .map(|req| req.client_id);
-            let infos = shard.list_all_clients().await;
-            #[allow(clippy::cast_possible_truncation)]
-            let found = target.and_then(|id| infos.iter().find(|info| info.client_id as u32 == id));
-            // The SDK decodes an empty body as `None` (client not found).
-            let bytes = found.map_or_else(Bytes::new, |info| {
-                ClientDetailsResponse {
-                    client: connected_client_to_response(info),
-                    consumer_groups: Vec::new(),
-                }
-                .to_bytes()
-            });
-            send_non_replicated_bytes(shard, &request, transport_client_id, bytes, "get_client")
-                .await;
-        }
-        _ => match build_non_replicated_response(shard, code, request_body(&request)) {
-            Ok(response) => {
-                let commit = current_metadata_commit(shard);
-                let reply = response.into_reply(
-                    request.header(),
-                    request.header().client,
-                    request.header().session,
-                    commit,
-                );
-                if let Err(error) = shard
-                    .bus
-                    .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-                    .await
-                {
-                    warn!(
-                        transport_client_id,
-                        code,
-                        error = %error,
-                        "failed to send non-replicated VSR reply"
-                    );
-                }
-            }
-            Err(error) => {
-                warn!(
-                    transport_client_id,
-                    code,
-                    error = %error,
-                    "dropping unsupported non-replicated VSR request"
-                );
-            }
-        },
-    }
-}
-
-/// Build the `get_me` reply for the requesting connection, sourced
-/// entirely from the per-shard [`SessionManager`] (`user_id`, transport
-/// kind, peer address) -- no message-bus lookup. `consumer_groups` is
-/// empty (see [`client_record_to_response`]).
-fn build_get_me_response(
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-) -> ClientDetailsResponse {
-    let client = sessions
-        .borrow()
-        .client_record(transport_client_id)
-        .map_or_else(
-            || {
-                // No session record (shouldn't happen on an auth-gated
-                // read). Report the connection id with the "no user"
-                // sentinel + TCP default rather than impersonating root
-                // (user id 0 is a real user; server-ng is 0-based).
-                #[allow(clippy::cast_possible_truncation)]
-                ClientResponse {
-                    client_id: transport_client_id as u32,
-                    user_id: u32::MAX,
-                    transport: 1,
-                    address: String::new(),
-                    consumer_groups_count: 0,
-                }
-            },
-            |record| client_record_to_response(&record),
-        );
-    ClientDetailsResponse {
-        client,
-        consumer_groups: Vec::new(),
-    }
-}
-
-/// Convert a [`ClientRecord`] (per-shard session state) into the wire
-/// [`ClientResponse`]. Shared by `get_me` and the `get_clients` gather.
-///
-/// TODO(consumer-group-membership): `consumer_groups_count` is always 0
-/// -- server-ng does not yet track which consumer groups a connection
-/// joined. Populate once the partition-reconciliation / consumer-group
-/// rebalancing work lands; until then `get_me` / `get_clients` report no
-/// memberships.
-fn client_record_to_response(record: &ClientRecord) -> ClientResponse {
-    // The transport client id is a u128 `(shard << 112) | seq`; the wire
-    // `client_id` is the u32 seq tail.
-    #[allow(clippy::cast_possible_truncation)]
-    ClientResponse {
-        client_id: record.client_id as u32,
-        user_id: record.user_id.unwrap_or(u32::MAX),
-        transport: transport_kind_to_wire(record.transport),
-        address: record.address.to_string(),
-        consumer_groups_count: 0,
-    }
-}
-
-/// Convert a gathered [`ConnectedClientInfo`] (one shard's view of a
-/// connected client) into the wire [`ClientResponse`] for `get_clients` /
-/// `get_client`. Same `consumer_groups_count: 0` caveat as
-/// [`client_record_to_response`].
-fn connected_client_to_response(info: &ConnectedClientInfo) -> ClientResponse {
-    #[allow(clippy::cast_possible_truncation)]
-    ClientResponse {
-        client_id: info.client_id as u32,
-        user_id: info.user_id.unwrap_or(u32::MAX),
-        transport: transport_kind_to_wire(info.transport),
-        address: info.address.to_string(),
-        consumer_groups_count: 0,
-    }
-}
-
-/// Send a non-replicated reply body to a client, stamping the current
-/// metadata commit. Shared by the `get_me` / `get_clients` / `get_client`
-/// arms.
-#[allow(clippy::future_not_send)]
-async fn send_non_replicated_bytes(
-    shard: &Rc<ServerNgShard>,
-    request: &Message<RequestHeader>,
-    transport_client_id: u128,
-    bytes: Bytes,
-    label: &'static str,
-) {
-    let commit = current_metadata_commit(shard);
-    let reply = NonReplicatedResponse::Bytes(bytes).into_reply(
-        request.header(),
-        request.header().client,
-        request.header().session,
-        commit,
-    );
-    if let Err(error) = shard
-        .bus
-        .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-        .await
-    {
-        warn!(transport_client_id, label, error = %error, "failed to send non-replicated reply");
-    }
-}
-
-/// Map the transport kind to the legacy wire discriminant
-/// (`1=TCP, 2=QUIC, 4=WebSocket`); TLS variants report their base
-/// transport. `ClientTransportKind` is `#[non_exhaustive]`, so any other
-/// (TCP, TCP-TLS, or a future) variant falls back to TCP.
-const fn transport_kind_to_wire(kind: ClientTransportKind) -> u8 {
-    match kind {
-        ClientTransportKind::Quic => 2,
-        ClientTransportKind::Ws | ClientTransportKind::Wss => 4,
-        _ => 1,
-    }
-}
-
-/// Run the consensus `Register` proposal on the metadata owner (shard 0)
-/// and return the committed session.
-///
-/// Credential verification and session binding stay on the calling (home)
-/// shard -- only this consensus step must execute where the metadata
-/// consensus group lives. On shard 0 it calls in-process directly; on a
-/// peer it forwards a [`shard::MetadataSubmit`] to shard 0 and awaits the
-/// committed op. A dropped reply (shard-0 inbox full / shutdown) maps to a
-/// transient `Canceled`, which the caller wraps so the SDK replays.
-#[allow(clippy::future_not_send)]
-async fn submit_register_on_owner(
-    shard: &Rc<ServerNgShard>,
-    vsr_client_id: u128,
-) -> Result<u64, RegisterSubmitError> {
-    if shard.id == 0 {
-        return shard
-            .plane
-            .metadata()
-            .submit_register_in_process(vsr_client_id)
-            .await;
-    }
-    let (reply, rx) = shard::channel::<Option<u64>>(1);
-    shard.forward_metadata_submit(shard::MetadataSubmit::Register {
-        vsr_client_id,
-        reply,
-    });
-    match rx.recv().await {
-        Ok(Some(session)) => Ok(session),
-        _ => Err(RegisterSubmitError::Canceled),
-    }
-}
-
-/// Logout counterpart of [`submit_register_on_owner`].
-#[allow(clippy::future_not_send)]
-async fn submit_logout_on_owner(
-    shard: &Rc<ServerNgShard>,
-    vsr_client_id: u128,
-    session: u64,
-    request: u64,
-) -> Result<u64, RegisterSubmitError> {
-    if shard.id == 0 {
-        return shard
-            .plane
-            .metadata()
-            .submit_logout_in_process(vsr_client_id, session, request)
-            .await;
-    }
-    let (reply, rx) = shard::channel::<Option<u64>>(1);
-    shard.forward_metadata_submit(shard::MetadataSubmit::Logout {
-        vsr_client_id,
-        session,
-        request,
-        reply,
-    });
-    match rx.recv().await {
-        Ok(Some(commit)) => Ok(commit),
-        _ => Err(RegisterSubmitError::Canceled),
-    }
-}
-
-/// Submit a replicated client request to the metadata owner (shard 0) and
-/// return the committed reply.
-///
-/// The metadata consensus group lives on shard 0, but the connection lives
-/// on the home shard (this shard). Run consensus where it belongs and bring
-/// the committed reply back here so the caller can write it to the
-/// originating socket -- shard 0 cannot route the reply by the consensus
-/// `client` id (it's the VSR id, not the transport/home-shard-encoding id).
-/// `None` = transient submit failure (SDK read-timeout replays).
-#[allow(clippy::future_not_send)]
-async fn submit_client_request_on_owner(
-    shard: &Rc<ServerNgShard>,
-    request: Message<RequestHeader>,
-) -> Option<Message<GenericHeader>> {
-    if shard.id == 0 {
-        return shard
-            .plane
-            .metadata()
-            .submit_request_in_process(request)
-            .await
-            .ok()
-            .map(server_common::Message::into_generic);
-    }
-    let (reply, rx) = shard::channel::<Option<Message<GenericHeader>>>(1);
-    shard.forward_metadata_submit(shard::MetadataSubmit::ClientRequest {
-        request: request.into_generic(),
-        reply,
-    });
-    rx.recv().await.ok().flatten()
-}
-
-#[allow(clippy::future_not_send)]
-async fn handle_logout_request(
-    shard: &Rc<ServerNgShard>,
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-    request: Message<RequestHeader>,
-) {
-    let Some((vsr_client_id, session)) = sessions.borrow().get_session(transport_client_id) else {
-        warn!(
-            transport_client_id,
-            "dropping logout for unbound VSR session"
-        );
-        return;
-    };
-
-    let request_id = request.header().request;
-    let commit = match submit_logout_on_owner(shard, vsr_client_id, session, request_id).await {
-        Ok(commit) => commit,
-        Err(error) => {
-            warn!(transport_client_id, error = %error, "logout/unregister failed");
-            return;
-        }
-    };
-
-    sessions.borrow_mut().remove_connection(transport_client_id);
-
-    let reply = build_empty_reply(request.header(), vsr_client_id, session, commit);
-    if let Err(error) = shard
-        .bus
-        .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-        .await
-    {
-        warn!(
-            transport_client_id,
-            error = %error,
-            "failed to send logout reply"
-        );
-    }
-}
-
-fn ensure_transport_connection(
-    shard: &Rc<ServerNgShard>,
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-) {
-    let Some(meta) = shard.bus.client_meta(transport_client_id) else {
-        return;
-    };
-    sessions
-        .borrow_mut()
-        .ensure_connection(transport_client_id, meta.peer_addr, meta.transport);
-}
-
-fn maybe_rewrite_pat_request(
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-    request: Message<RequestHeader>,
-) -> Result<(Message<RequestHeader>, Option<String>), IggyError> {
-    let operation = request.header().operation;
-    let user_id = match operation {
-        Operation::CreatePersonalAccessToken | Operation::DeletePersonalAccessToken => sessions
-            .borrow()
-            .get_user_id(transport_client_id)
-            .ok_or(IggyError::Unauthenticated)?,
-        _ => return Ok((request, None)),
-    };
-
-    let body = request_body(&request);
-    let mut raw_token = None;
-    let rewritten = match operation {
-        Operation::CreatePersonalAccessToken => {
-            let wire = WireCreatePersonalAccessTokenRequest::decode_from(body)
-                .map_err(|_| IggyError::InvalidCommand)?;
-            // Primary mints the raw token + hash here and ships the hash
-            // through consensus. Replicas decode the hash directly. Doing
-            // this inside `CreatePersonalAccessTokenRequest::apply` would
-            // call `ring::rand` per-replica and diverge state. The raw token
-            // is returned to this client only (see `handle_client_request`).
-            let (raw, token_hash) = mint_pat_raw_and_hash();
-            raw_token = Some(raw);
-            ReplicatedCreatePersonalAccessTokenRequest {
-                user_id,
-                name: wire.name,
-                expiry: wire.expiry,
-                token_hash,
-            }
-            .to_bytes()
-        }
-        Operation::DeletePersonalAccessToken => {
-            let wire = WireDeletePersonalAccessTokenRequest::decode_from(body)
-                .map_err(|_| IggyError::InvalidCommand)?;
-            ReplicatedDeletePersonalAccessTokenRequest {
-                user_id,
-                name: wire.name,
-            }
-            .to_bytes()
-        }
-        _ => unreachable!(),
-    };
-
-    Ok((rewrite_request_body(&request, &rewritten)?, raw_token))
-}
-
-/// Mints a fresh PAT and returns the raw token plus its hex-encoded SHA-256
-/// hash (64 bytes ASCII). Only the hash is replicated; the raw token is
-/// returned to the minting client by the home shard (it cannot be reproduced
-/// by the deterministic `apply` running on every replica).
-fn mint_pat_raw_and_hash() -> (String, [u8; 64]) {
-    let (raw, hash) = iggy_common::PersonalAccessToken::mint_raw_and_hash();
-    let bytes = hash.as_bytes();
-    let mut out = [0u8; 64];
-    let len = bytes.len().min(64);
-    out[..len].copy_from_slice(&bytes[..len]);
-    (raw, out)
-}
-
-fn rewrite_request_body(
-    request: &Message<RequestHeader>,
-    body: &Bytes,
-) -> Result<Message<RequestHeader>, IggyError> {
-    let total_size = std::mem::size_of::<RequestHeader>()
-        .checked_add(body.len())
-        .ok_or(IggyError::InvalidConfiguration)?;
-    let size = u32::try_from(total_size).map_err(|_| IggyError::InvalidConfiguration)?;
-    let mut rewritten = Message::<RequestHeader>::new(total_size);
-    let header = bytemuck::checked::try_from_bytes_mut::<RequestHeader>(
-        &mut rewritten.as_mut_slice()[..std::mem::size_of::<RequestHeader>()],
-    )
-    .expect("zeroed bytes are a valid request header");
-    *header = *request.header();
-    header.size = size;
-    rewritten.as_mut_slice()[std::mem::size_of::<RequestHeader>()..].copy_from_slice(body);
-    // TODO(vsr): the body changed but `request_checksum` / `checksum` /
-    // `checksum_body` were copied verbatim from the original header. Safe
-    // today because the SDK initializes `request_checksum` to 0 and the
-    // server does not validate it; the moment integrity checking lands,
-    // recompute these here (or zero them and re-sign in a follow-up step).
-    Ok(rewritten)
-}
-
-#[allow(clippy::future_not_send)]
-async fn handle_login_register_request(
-    shard: &Rc<ServerNgShard>,
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-    request: Message<RequestHeader>,
-) {
-    let body = request_body(&request);
-    let vsr_client_id = request.header().client;
-
-    if let Ok(wire_request) = LoginRegisterRequest::decode_from(body) {
-        match verify_login_credentials(
-            shard,
-            wire_request.username.as_str(),
-            wire_request.password.expose_secret(),
-        ) {
-            Ok(user_id) => {
-                if let Err(error) = complete_login_register(
-                    shard,
-                    sessions,
-                    transport_client_id,
-                    vsr_client_id,
-                    request.header(),
-                    user_id,
-                )
-                .await
-                {
-                    warn!(transport_client_id, error = %error, "login/register failed");
-                    surface_login_failure(shard, transport_client_id, request.header(), &error)
-                        .await;
-                }
-                return;
-            }
-            Err(LoginRegisterError::InvalidCredentials) => {
-                // Fall through to PAT attempt so a credential payload that
-                // collides with a valid PAT payload shape still gets a
-                // chance; if PAT also rejects, the final fall-through emits
-                // the empty-reply failure path below.
-            }
-            Err(error) => {
-                warn!(transport_client_id, error = %error, "login/register failed");
-                surface_login_failure(shard, transport_client_id, request.header(), &error).await;
-                return;
-            }
-        }
-    }
-
-    if let Ok(wire_request) = LoginRegisterWithPatRequest::decode_from(body) {
-        match verify_pat_credentials(shard, wire_request.token.expose_secret()) {
-            Ok(user_id) => {
-                if let Err(error) = complete_login_register(
-                    shard,
-                    sessions,
-                    transport_client_id,
-                    vsr_client_id,
-                    request.header(),
-                    user_id,
-                )
-                .await
-                {
-                    warn!(
-                        transport_client_id,
-                        error = %error,
-                        "login/register with PAT failed"
-                    );
-                    surface_login_failure(shard, transport_client_id, request.header(), &error)
-                        .await;
-                }
-                return;
-            }
-            Err(error) => {
-                warn!(
-                    transport_client_id,
-                    error = %error,
-                    "login/register with PAT failed"
-                );
-                surface_login_failure(shard, transport_client_id, request.header(), &error).await;
-                return;
-            }
-        }
-    }
-
-    warn!(
-        transport_client_id,
-        "dropping register request with unsupported payload shape"
-    );
-    send_login_failure_reply(shard, transport_client_id, request.header()).await;
-}
-
-/// Decide whether a failed login/register gets a terminal reply or silence.
-///
-/// `Transient` / `InvalidClientId` are `NotEvictable` (see
-/// [`LoginRegisterError`]'s `TryFrom` for `EvictionReason`): the cluster
-/// could not commit *right now* (e.g. a freshly booted primary still
-/// catching up, or a cross-shard submit canceled). Staying silent lets the
-/// SDK read-timeout replay -- a later attempt lands once the primary is
-/// caught up. Replying empty here would instead surface as a hard
-/// `InvalidFormat` decode failure and break the replay.
-///
-/// Terminal auth errors (`InvalidCredentials` / `InvalidToken` /
-/// `UserInactive` / `Session`) map to an `EvictionReason`, so we fast-fail
-/// with an empty reply rather than make the client wait for a timeout.
-/// (TODO: ship a typed `Eviction` frame once the SDK eviction decoder lands
-/// on every transport.)
-#[allow(clippy::future_not_send)]
-async fn surface_login_failure(
-    shard: &Rc<ServerNgShard>,
-    transport_client_id: u128,
-    request_header: &RequestHeader,
-    error: &LoginRegisterError,
-) {
-    if EvictionReason::try_from(error).is_ok() {
-        send_login_failure_reply(shard, transport_client_id, request_header).await;
-    }
-}
-
-/// Empty Reply on a terminal failed Register. The SDK only decodes
-/// `Command2::Reply`; an empty body fails `LoginRegisterResponse` decoding
-/// fast instead of hanging until the socket read timeout. Only call for
-/// terminal errors -- see [`surface_login_failure`].
-#[allow(clippy::future_not_send)]
-async fn send_login_failure_reply(
-    shard: &Rc<ServerNgShard>,
-    transport_client_id: u128,
-    request_header: &RequestHeader,
-) {
-    let commit = current_metadata_commit(shard);
-    let reply = build_empty_reply(request_header, transport_client_id, 0, commit);
-    if let Err(error) = shard
-        .bus
-        .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-        .await
-    {
-        warn!(
-            transport_client_id,
-            error = %error,
-            "failed to send login-failure reply"
-        );
-    }
-}
-
-fn request_body(request: &Message<RequestHeader>) -> &[u8] {
-    &request.as_slice()[std::mem::size_of::<RequestHeader>()..request.header().size as usize]
-}
-
-fn resolve_partition_request_namespace(
-    shard: &Rc<ServerNgShard>,
-    operation: Operation,
-    body: &[u8],
-) -> Result<u64, IggyError> {
-    let namespace = match operation {
-        Operation::SendMessages => {
-            if body.len() < 4 {
-                return Err(IggyError::InvalidCommand);
-            }
-            let metadata_length = u32::from_le_bytes(
-                body[..4]
-                    .try_into()
-                    .map_err(|_| IggyError::InvalidNumberEncoding)?,
-            ) as usize;
-            if body.len() < 4 + metadata_length {
-                return Err(IggyError::InvalidCommand);
-            }
-            let header = SendMessagesHeader::decode_from(&body[4..4 + metadata_length])
-                .map_err(|_| IggyError::InvalidCommand)?;
-            resolve_send_messages_namespace(shard, &header)?
-        }
-        Operation::StoreConsumerOffset => {
-            let request = StoreConsumerOffsetRequest::decode_from(body)
-                .map_err(|_| IggyError::InvalidCommand)?;
-            resolve_partition_namespace(
-                shard,
-                &request.stream_id,
-                &request.topic_id,
-                request.partition_id,
-            )?
-        }
-        Operation::DeleteConsumerOffset => {
-            let request = DeleteConsumerOffsetRequest::decode_from(body)
-                .map_err(|_| IggyError::InvalidCommand)?;
-            resolve_partition_namespace(
-                shard,
-                &request.stream_id,
-                &request.topic_id,
-                request.partition_id,
-            )?
-        }
-        Operation::StoreConsumerOffset2 => {
-            let request = StoreConsumerOffset2Request::decode_from(body)
-                .map_err(|_| IggyError::InvalidCommand)?;
-            resolve_partition_namespace(
-                shard,
-                &request.stream_id,
-                &request.topic_id,
-                request.partition_id,
-            )?
-        }
-        Operation::DeleteConsumerOffset2 => {
-            let request = DeleteConsumerOffset2Request::decode_from(body)
-                .map_err(|_| IggyError::InvalidCommand)?;
-            resolve_partition_namespace(
-                shard,
-                &request.stream_id,
-                &request.topic_id,
-                request.partition_id,
-            )?
-        }
-        Operation::DeleteSegments => {
-            let request =
-                DeleteSegmentsRequest::decode_from(body).map_err(|_| IggyError::InvalidCommand)?;
-            resolve_partition_namespace(
-                shard,
-                &request.stream_id,
-                &request.topic_id,
-                Some(request.partition_id),
-            )?
-        }
-        _ => return Err(IggyError::FeatureUnavailable),
-    };
-    Ok(namespace.inner())
-}
-
-fn resolve_send_messages_namespace(
-    shard: &Rc<ServerNgShard>,
-    header: &SendMessagesHeader,
-) -> Result<IggyNamespace, IggyError> {
-    let WirePartitioning::PartitionId(partition_id) = header.partitioning else {
-        return Err(IggyError::FeatureUnavailable);
-    };
-    resolve_partition_namespace(
-        shard,
-        &header.stream_id,
-        &header.topic_id,
-        Some(partition_id),
-    )
-}
-
-fn resolve_partition_namespace(
-    shard: &Rc<ServerNgShard>,
-    stream_id: &WireIdentifier,
-    topic_id: &WireIdentifier,
-    partition_id: Option<u32>,
-) -> Result<IggyNamespace, IggyError> {
-    let partition_id = partition_id.ok_or(IggyError::InvalidIdentifier)?;
-    shard
-        .plane
-        .metadata()
-        .mux_stm
-        .streams()
-        .namespace_from_partition(stream_id, topic_id, partition_id)
-        .ok_or(IggyError::InvalidIdentifier)
-}
-
-fn build_non_replicated_response(
-    shard: &Rc<ServerNgShard>,
-    code: u32,
-    body: &[u8],
-) -> Result<NonReplicatedResponse, IggyError> {
-    match code {
-        GET_CLUSTER_METADATA_CODE => Ok(NonReplicatedResponse::Bytes(
-            build_cluster_metadata_response().to_bytes(),
-        )),
-        GET_STATS_CODE => Ok(NonReplicatedResponse::Bytes(
-            build_stats_response(shard)?.to_bytes(),
-        )),
-        GET_STREAM_CODE => {
-            let request =
-                GetStreamRequest::decode_from(body).map_err(|_| IggyError::InvalidCommand)?;
-            build_get_stream_response(shard, &request.stream_id).map(|response| {
-                response.map_or(NonReplicatedResponse::Empty, |response| {
-                    NonReplicatedResponse::Bytes(response.to_bytes())
-                })
-            })
-        }
-        GET_STREAMS_CODE => {
-            let _ = GetStreamsRequest::decode_from(body).map_err(|_| IggyError::InvalidCommand)?;
-            Ok(NonReplicatedResponse::Bytes(
-                build_get_streams_response(shard)?.to_bytes(),
-            ))
-        }
-        GET_TOPIC_CODE => {
-            let request =
-                GetTopicRequest::decode_from(body).map_err(|_| IggyError::InvalidCommand)?;
-            build_get_topic_response(shard, &request.stream_id, &request.topic_id).map(|response| {
-                response.map_or(NonReplicatedResponse::Empty, |response| {
-                    NonReplicatedResponse::Bytes(response.to_bytes())
-                })
-            })
-        }
-        GET_TOPICS_CODE => {
-            let request =
-                GetTopicsRequest::decode_from(body).map_err(|_| IggyError::InvalidCommand)?;
-            Ok(NonReplicatedResponse::Bytes(
-                build_get_topics_response(shard, &request.stream_id)?.to_bytes(),
-            ))
-        }
-        GET_USERS_CODE => Ok(NonReplicatedResponse::Bytes(
-            build_get_users_response(shard)?.to_bytes(),
-        )),
-        GET_USER_CODE => {
-            let request =
-                GetUserRequest::decode_from(body).map_err(|_| IggyError::InvalidCommand)?;
-            build_get_user_response(shard, &request.user_id).map(|response| {
-                response.map_or(NonReplicatedResponse::Empty, |response| {
-                    NonReplicatedResponse::Bytes(response.to_bytes())
-                })
-            })
-        }
-        POLL_MESSAGES_CODE => {
-            let request =
-                PollMessagesRequest::decode_from(body).map_err(|_| IggyError::InvalidCommand)?;
-            Ok(NonReplicatedResponse::EmptyPolledMessages {
-                partition_id: request.partition_id.unwrap_or(0),
-            })
-        }
-        _ => match iggy_binary_protocol::dispatch::lookup_command(code) {
-            Some(meta) if !meta.is_replicated() => Ok(NonReplicatedResponse::Empty),
-            Some(_) => Err(IggyError::FeatureUnavailable),
-            None => Err(IggyError::InvalidCommand),
-        },
-    }
-}
-
-fn build_cluster_metadata_response() -> ClusterMetadataResponse {
-    ClusterMetadataResponse {
-        name: "server-ng".to_owned(),
-        nodes: vec![ClusterNodeResponse {
-            name: "node-0".to_owned(),
-            ip: "127.0.0.1".to_owned(),
-            tcp_port: 0,
-            quic_port: 0,
-            http_port: 0,
-            websocket_port: 0,
-            role: 0,
-            status: 0,
-        }],
-    }
-}
-
-fn build_stats_response(shard: &Rc<ServerNgShard>) -> Result<StatsResponse, IggyError> {
-    let (streams_count, topics_count, partitions_count, messages_size_bytes, messages_count) =
-        shard
-            .plane
-            .metadata()
-            .mux_stm
-            .streams()
-            .read(|streams| -> Result<_, IggyError> {
-                let mut topics_count = 0u32;
-                let mut partitions_count = 0u32;
-                let mut messages_size_bytes = 0u64;
-                let mut messages_count = 0u64;
-                for (_, stream) in &streams.items {
-                    topics_count = topics_count.saturating_add(usize_to_u32(stream.topics.len())?);
-                    messages_size_bytes =
-                        messages_size_bytes.saturating_add(stream.stats.size_bytes_inconsistent());
-                    messages_count =
-                        messages_count.saturating_add(stream.stats.messages_count_inconsistent());
-                    for (_, topic) in &stream.topics {
-                        partitions_count =
-                            partitions_count.saturating_add(usize_to_u32(topic.partitions.len())?);
-                    }
-                }
-                Ok((
-                    usize_to_u32(streams.items.len())?,
-                    topics_count,
-                    partitions_count,
-                    messages_size_bytes,
-                    messages_count,
-                ))
-            })?;
-    let consumer_groups_count = shard
-        .plane
-        .metadata()
-        .mux_stm
-        .consumer_groups()
-        .read(|groups| usize_to_u32(groups.items.len()))?;
-
-    Ok(StatsResponse {
-        process_id: std::process::id(),
-        cpu_usage: 0.0,
-        total_cpu_usage: 0.0,
-        memory_usage: 0,
-        total_memory: 0,
-        available_memory: 0,
-        run_time: 0,
-        start_time: 0,
-        read_bytes: 0,
-        written_bytes: 0,
-        messages_size_bytes,
-        streams_count,
-        topics_count,
-        partitions_count,
-        segments_count: 0,
-        messages_count,
-        clients_count: 0,
-        consumer_groups_count,
-        hostname: "unknown_hostname".to_owned(),
-        os_name: "unknown_os_name".to_owned(),
-        os_version: "unknown_os_version".to_owned(),
-        kernel_version: "unknown_kernel_version".to_owned(),
-        iggy_server_version: server::VERSION.to_owned(),
-        iggy_server_semver: server::SEMANTIC_VERSION.get_numeric_version().ok(),
-        cache_metrics: Vec::new(),
-        threads_count: 0,
-        free_disk_space: 0,
-        total_disk_space: 0,
-    })
-}
-
-fn build_get_stream_response(
-    shard: &Rc<ServerNgShard>,
-    stream_id: &WireIdentifier,
-) -> Result<Option<GetStreamResponse>, IggyError> {
-    shard.plane.metadata().mux_stm.streams().read(|streams| {
-        let Some(stream_id) = resolve_stream_id(streams, stream_id) else {
-            return Ok(None);
-        };
-        let stream = streams
-            .items
-            .get(stream_id)
-            .ok_or(IggyError::InvalidIdentifier)?;
-        Ok(Some(GetStreamResponse {
-            stream: stream_response(stream)?,
-            topics: stream
-                .topics
-                .iter()
-                .map(|(_, topic)| topic_header(topic))
-                .collect::<Result<Vec<_>, _>>()?,
-        }))
-    })
-}
-
-fn build_get_streams_response(shard: &Rc<ServerNgShard>) -> Result<GetStreamsResponse, IggyError> {
-    shard.plane.metadata().mux_stm.streams().read(|streams| {
-        streams
-            .items
-            .iter()
-            .map(|(_, stream)| stream_response(stream))
-            .collect::<Result<Vec<_>, _>>()
-            .map(|streams| GetStreamsResponse { streams })
-    })
-}
-
-#[allow(clippy::cast_possible_truncation)]
-fn user_response(user: &metadata::stm::user::User) -> Result<UserResponse, IggyError> {
-    Ok(UserResponse {
-        id: user.id,
-        created_at: user.created_at.as_micros(),
-        status: user.status.as_code(),
-        username: WireName::new(user.username.as_ref()).map_err(|_| IggyError::InvalidFormat)?,
-    })
-}
-
-fn build_get_users_response(shard: &Rc<ServerNgShard>) -> Result<GetUsersResponse, IggyError> {
-    shard.plane.metadata().mux_stm.users().read(|users| {
-        users
-            .items
-            .iter()
-            .map(|(_, user)| user_response(user))
-            .collect::<Result<Vec<_>, _>>()
-            .map(|users| GetUsersResponse { users })
-    })
-}
-
-fn build_get_user_response(
-    shard: &Rc<ServerNgShard>,
-    user_id: &WireIdentifier,
-) -> Result<Option<UserDetailsResponse>, IggyError> {
-    shard.plane.metadata().mux_stm.users().read(|users| {
-        let resolved = match user_id {
-            WireIdentifier::Numeric(id) => {
-                let id = *id as usize;
-                users.items.contains(id).then_some(id)
-            }
-            WireIdentifier::String(name) => users.index.get(name.as_str()).map(|&id| id as usize),
-        };
-        let Some(id) = resolved else {
-            return Ok(None);
-        };
-        let user = users.items.get(id).ok_or(IggyError::InvalidIdentifier)?;
-        Ok(Some(UserDetailsResponse {
-            user: user_response(user)?,
-            permissions: user
-                .permissions
-                .as_ref()
-                .map(|p| iggy_common::wire_conversions::permissions_to_wire(p)),
-        }))
-    })
-}
-
-fn build_get_topic_response(
-    shard: &Rc<ServerNgShard>,
-    stream_id: &WireIdentifier,
-    topic_id: &WireIdentifier,
-) -> Result<Option<GetTopicResponse>, IggyError> {
-    shard.plane.metadata().mux_stm.streams().read(|streams| {
-        let Some(stream_id) = resolve_stream_id(streams, stream_id) else {
-            return Ok(None);
-        };
-        let Some(topic_id) = resolve_topic_id(streams, stream_id, topic_id) else {
-            return Ok(None);
-        };
-        let stream = streams
-            .items
-            .get(stream_id)
-            .ok_or(IggyError::InvalidIdentifier)?;
-        let topic = stream
-            .topics
-            .get(topic_id)
-            .ok_or(IggyError::InvalidIdentifier)?;
-        Ok(Some(GetTopicResponse {
-            topic: topic_header(topic)?,
-            partitions: topic
-                .partitions
-                .iter()
-                .map(partition_response)
-                .collect::<Result<Vec<_>, _>>()?,
-        }))
-    })
-}
-
-fn build_get_topics_response(
-    shard: &Rc<ServerNgShard>,
-    stream_id: &WireIdentifier,
-) -> Result<GetTopicsResponse, IggyError> {
-    shard.plane.metadata().mux_stm.streams().read(|streams| {
-        let Some(stream_id) = resolve_stream_id(streams, stream_id) else {
-            return Ok(GetTopicsResponse { topics: Vec::new() });
-        };
-        let stream = streams
-            .items
-            .get(stream_id)
-            .ok_or(IggyError::InvalidIdentifier)?;
-        stream
-            .topics
-            .iter()
-            .map(|(_, topic)| topic_header(topic))
-            .collect::<Result<Vec<_>, _>>()
-            .map(|topics| GetTopicsResponse { topics })
-    })
-}
-
-fn resolve_stream_id(
-    streams: &metadata::stm::stream::StreamsInner,
-    identifier: &WireIdentifier,
-) -> Option<usize> {
-    match identifier {
-        WireIdentifier::Numeric(id) => {
-            let id = *id as usize;
-            streams.items.contains(id).then_some(id)
-        }
-        WireIdentifier::String(name) => streams.index.get(name.as_str()).copied(),
-    }
-}
-
-fn resolve_topic_id(
-    streams: &metadata::stm::stream::StreamsInner,
-    stream_id: usize,
-    identifier: &WireIdentifier,
-) -> Option<usize> {
-    let stream = streams.items.get(stream_id)?;
-    match identifier {
-        WireIdentifier::Numeric(id) => {
-            let id = *id as usize;
-            stream.topics.contains(id).then_some(id)
-        }
-        WireIdentifier::String(name) => stream.topic_index.get(name.as_str()).copied(),
-    }
-}
-
-fn stream_response(stream: &metadata::stm::stream::Stream) -> Result<StreamResponse, IggyError> {
-    Ok(StreamResponse {
-        id: usize_to_u32(stream.id)?,
-        created_at: stream.created_at.as_micros(),
-        topics_count: usize_to_u32(stream.topics.len())?,
-        size_bytes: stream.stats.size_bytes_inconsistent(),
-        messages_count: stream.stats.messages_count_inconsistent(),
-        name: WireName::new(stream.name.as_ref()).map_err(|_| IggyError::InvalidFormat)?,
-    })
-}
-
-fn topic_header(topic: &metadata::stm::stream::Topic) -> Result<StreamTopicHeader, IggyError> {
-    Ok(StreamTopicHeader {
-        id: usize_to_u32(topic.id)?,
-        created_at: topic.created_at.as_micros(),
-        partitions_count: usize_to_u32(topic.partitions.len())?,
-        message_expiry: u64::from(topic.message_expiry),
-        compression_algorithm: topic.compression_algorithm.as_code(),
-        max_topic_size: topic.max_topic_size.as_bytes_u64(),
-        replication_factor: topic.replication_factor,
-        size_bytes: topic.stats.size_bytes_inconsistent(),
-        messages_count: topic.stats.messages_count_inconsistent(),
-        name: WireName::new(topic.name.as_ref()).map_err(|_| IggyError::InvalidFormat)?,
-    })
-}
-
-fn partition_response(
-    partition: &metadata::stm::stream::Partition,
-) -> Result<PartitionResponse, IggyError> {
-    Ok(PartitionResponse {
-        id: usize_to_u32(partition.id)?,
-        created_at: partition.created_at.as_micros(),
-        segments_count: 0,
-        current_offset: 0,
-        size_bytes: 0,
-        messages_count: 0,
-    })
-}
-
-fn usize_to_u32(value: usize) -> Result<u32, IggyError> {
-    u32::try_from(value).map_err(|_| IggyError::InvalidIdentifier)
-}
-
-fn verify_login_credentials(
-    shard: &Rc<ServerNgShard>,
-    username: &str,
-    password: &str,
-) -> Result<u32, LoginRegisterError> {
-    shard.plane.metadata().mux_stm.users().read(|users| {
-        let Some(user_id) = users.index.get(username).copied() else {
-            return Err(LoginRegisterError::InvalidCredentials);
-        };
-        let Some(user) = users.items.get(user_id as usize) else {
-            return Err(LoginRegisterError::InvalidCredentials);
-        };
-        if user.status != UserStatus::Active {
-            return Err(LoginRegisterError::UserInactive);
-        }
-        if !crypto::verify_password(password, user.password_hash.as_ref()) {
-            return Err(LoginRegisterError::InvalidCredentials);
-        }
-        Ok(user.id)
-    })
-}
-
-fn verify_pat_credentials(
-    shard: &Rc<ServerNgShard>,
-    token: &str,
-) -> Result<u32, LoginRegisterError> {
-    let token_hash = PersonalAccessToken::hash_token(token);
-    let now = IggyTimestamp::now();
-    shard.plane.metadata().mux_stm.users().read(|users| {
-        let Some((user_id, token_name)) =
-            users.personal_access_token_index.get(token_hash.as_str())
-        else {
-            return Err(LoginRegisterError::InvalidToken);
-        };
-        let Some(pat) = users
-            .personal_access_tokens
-            .get(user_id)
-            .and_then(|tokens| tokens.get(token_name))
-        else {
-            return Err(LoginRegisterError::InvalidToken);
-        };
-        if pat.is_expired(now) {
-            return Err(LoginRegisterError::InvalidToken);
-        }
-        let Some(user) = users.items.get(*user_id as usize) else {
-            return Err(LoginRegisterError::InvalidToken);
-        };
-        if user.status != UserStatus::Active {
-            return Err(LoginRegisterError::UserInactive);
-        }
-        Ok(user.id)
-    })
-}
-
-#[allow(clippy::future_not_send)]
-async fn complete_login_register(
-    shard: &Rc<ServerNgShard>,
-    sessions: &Rc<RefCell<SessionManager>>,
-    transport_client_id: u128,
-    vsr_client_id: u128,
-    request_header: &RequestHeader,
-    user_id: u32,
-) -> Result<(), LoginRegisterError> {
-    let existing_session = {
-        let sessions = sessions.borrow();
-        sessions
-            .get_session(transport_client_id)
-            .map(|(_, session)| session)
-    };
-    if let Some(session) = existing_session {
-        let commit = current_metadata_commit(shard);
-        let reply =
-            build_login_register_reply(request_header, vsr_client_id, session, commit, user_id);
-        let _ = shard
-            .bus
-            .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-            .await;
-        return Ok(());
-    }
-
-    {
-        let mut sessions = sessions.borrow_mut();
-        sessions
-            .login(transport_client_id, user_id)
-            .map_err(LoginRegisterError::Session)?;
-    }
-
-    let session = match submit_register_on_owner(shard, vsr_client_id).await {
-        Ok(session) => session,
-        Err(error) => {
-            let _ = sessions
-                .borrow_mut()
-                .reset_to_connected(transport_client_id);
-            return Err(LoginRegisterError::Transient(error));
-        }
-    };
-
-    {
-        let mut sessions = sessions.borrow_mut();
-        if let Err(error) = sessions.bind_session(transport_client_id, vsr_client_id, session) {
-            // No local rollback: `submit_register_in_process` above has
-            // already committed cluster-wide. A local-only
-            // `remove_client_session` here would diverge peers (they retain
-            // the slot until they evict the client themselves). The
-            // transport-disconnect callback owns local cleanup once the
-            // socket closes.
-            return Err(LoginRegisterError::Session(error));
-        }
-    }
-
-    let commit = current_metadata_commit(shard);
-    let reply = build_login_register_reply(request_header, vsr_client_id, session, commit, user_id);
-    if let Err(error) = shard
-        .bus
-        .send_to_client(transport_client_id, reply.into_generic().into_frozen())
-        .await
-    {
-        warn!(
-            transport_client_id,
-            error = %error,
-            "failed to send login/register reply"
-        );
-    }
-
-    Ok(())
-}
-
 fn ensure_default_root_user(mux_stm: &ServerNgMuxStateMachine) {
     if !mux_stm.users().read(|users| users.items.is_empty()) {
         return;
@@ -3760,158 +2063,6 @@ fn validate_cluster_root_bootstrap(
         username_env: IGGY_ROOT_USERNAME_ENV,
         password_env: IGGY_ROOT_PASSWORD_ENV,
     })
-}
-
-enum NonReplicatedResponse {
-    Empty,
-    EmptyPolledMessages { partition_id: u32 },
-    Bytes(Bytes),
-}
-
-impl NonReplicatedResponse {
-    fn into_reply(
-        self,
-        request_header: &RequestHeader,
-        client_id: u128,
-        session: u64,
-        commit: u64,
-    ) -> Message<ReplyHeader> {
-        match self {
-            Self::Empty => build_empty_reply(request_header, client_id, session, commit),
-            Self::EmptyPolledMessages { partition_id } => {
-                build_reply_with_body(request_header, client_id, session, commit, 16, |body| {
-                    body[..4].copy_from_slice(&partition_id.to_le_bytes());
-                    body[4..12].copy_from_slice(&0u64.to_le_bytes());
-                    body[12..16].copy_from_slice(&0u32.to_le_bytes());
-                })
-            }
-            Self::Bytes(body) => {
-                build_reply_from_bytes(request_header, client_id, session, commit, &body)
-            }
-        }
-    }
-}
-
-fn build_empty_reply(
-    request_header: &RequestHeader,
-    client_id: u128,
-    session: u64,
-    commit: u64,
-) -> Message<ReplyHeader> {
-    build_reply_with_body(request_header, client_id, session, commit, 0, |_| {})
-}
-
-fn build_login_register_reply(
-    request_header: &RequestHeader,
-    client_id: u128,
-    session: u64,
-    commit: u64,
-    user_id: u32,
-) -> Message<ReplyHeader> {
-    build_reply_with_body(request_header, client_id, session, commit, 12, |body| {
-        body[..4].copy_from_slice(&user_id.to_le_bytes());
-        body[4..12].copy_from_slice(&session.to_le_bytes());
-    })
-}
-
-fn build_reply_from_bytes(
-    request_header: &RequestHeader,
-    client_id: u128,
-    session: u64,
-    commit: u64,
-    body: &Bytes,
-) -> Message<ReplyHeader> {
-    build_reply_with_body(
-        request_header,
-        client_id,
-        session,
-        commit,
-        body.len(),
-        |out| out.copy_from_slice(body),
-    )
-}
-
-/// If a raw PAT token was minted (`CreatePersonalAccessToken`), replace the
-/// committed reply -- whose body is empty because the raw token never entered
-/// consensus -- with a `RawPersonalAccessTokenResponse`, reusing the confirmed
-/// commit position from the committed reply. Otherwise the committed reply
-/// passes through unchanged.
-fn build_raw_pat_reply(
-    request_header: &RequestHeader,
-    committed: Message<GenericHeader>,
-    raw_token: Option<String>,
-) -> Result<Message<GenericHeader>, IggyError> {
-    let Some(raw) = raw_token else {
-        return Ok(committed);
-    };
-    let header_len = std::mem::size_of::<ReplyHeader>();
-    let committed_header =
-        bytemuck::checked::try_from_bytes::<ReplyHeader>(&committed.as_slice()[..header_len])
-            .map_err(|_| IggyError::InvalidFormat)?;
-    let commit = committed_header.commit;
-    let token = WireName::new(raw.as_str()).map_err(|_| IggyError::InvalidFormat)?;
-    let body = RawPersonalAccessTokenResponse { token }.to_bytes();
-    let reply = build_reply_from_bytes(
-        request_header,
-        request_header.client,
-        request_header.session,
-        commit,
-        &body,
-    );
-    Ok(reply.into_generic())
-}
-
-fn build_reply_with_body(
-    request_header: &RequestHeader,
-    client_id: u128,
-    session: u64,
-    commit: u64,
-    body_len: usize,
-    write_body: impl FnOnce(&mut [u8]),
-) -> Message<ReplyHeader> {
-    let header_len = std::mem::size_of::<ReplyHeader>();
-    let total_size = header_len + body_len;
-    let mut reply = Message::<ReplyHeader>::new(total_size);
-    let header_size = u32::try_from(total_size).expect("reply size must fit into u32");
-    let header = bytemuck::checked::try_from_bytes_mut::<ReplyHeader>(
-        &mut reply.as_mut_slice()[..header_len],
-    )
-    .expect("zeroed bytes are valid");
-    *header = ReplyHeader {
-        cluster: request_header.cluster,
-        size: header_size,
-        view: request_header.view,
-        release: request_header.release,
-        command: Command2::Reply,
-        replica: request_header.replica,
-        request_checksum: request_header.request_checksum,
-        client: client_id,
-        op: session,
-        commit,
-        timestamp: request_header.timestamp,
-        request: request_header.request,
-        operation: request_header.operation,
-        namespace: request_header.namespace,
-        ..Default::default()
-    };
-    write_body(&mut reply.as_mut_slice()[header_len..total_size]);
-    reply
-}
-
-fn current_metadata_commit(shard: &Rc<ServerNgShard>) -> u64 {
-    shard
-        .plane
-        .metadata()
-        .consensus
-        .as_ref()
-        .map_or(0, VsrConsensus::commit_max)
-}
-
-fn upgrade_shard_handle(shard_handle: &ServerNgShardHandle) -> Option<Rc<ServerNgShard>> {
-    shard_handle
-        .borrow()
-        .as_ref()
-        .and_then(std::rc::Weak::upgrade)
 }
 
 /// Replica accept callback that ships every inbound connection through
